transaction_service/requirements.txt
Django>=4.2,<5.0
djangorestframework
drf-spectacular
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise
gunicorn
python-dotenv
requests
openai
